requirements.txt
asgiref==3.7.2
Cython==3.0.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
greenlet==2.0.2
joblib==1.3.2
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
patsy==0.5.3
pmdarima==2.0.3
PyJWT==2.8.0
python-dateutil==2.8.2
tenacity==8.2.3
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wheel==0.41.2
XlsxWriter==3.1.4
sqlalchemy
plotly
prophet